0.2 release

+++ retodo/commands/task.py
@@ -1,9 +1,5 @@
 #!/usr/bin/python3
 
-# PyTodo
-# Simple CLI to-do
-# v0.0.1
-
 import json
 import click
 import questionary
@@ -15,15 +11,33 @@ from rich.table import Table
 from rich.console import Console
 from rich import box
 
-DB_FILE_TASKS = os.environ['HOME'] + '/.pytodo/tasks.json'
-DB_FILE_CATEGORY = os.environ['HOME'] + '/.pytodo/category.json'
+DB_FILE_TASKS = os.environ['HOME'] + '/.retodo/tasks.json'
+DB_FILE_CATEGORY = os.environ['HOME'] + '/.retodo/category.json'
 
 
-@click.group()
-def cli():
-    pass
+# click class for alias
+class AliasedGroup(click.Group):
+
+    def get_command(self, ctx, cmd_name):
+        rv = click.Group.get_command(self, ctx, cmd_name)
+        if rv is not None:
+            return rv
+        matches = [x for x in self.list_commands(ctx)
+                   if x.startswith(cmd_name)]
+        if not matches:
+            return None
+        elif len(matches) == 1:
+            return click.Group.get_command(self, ctx, matches[0])
+        ctx.fail('Too many matches: %s' % ', '.join(sorted(matches)))
 
 
+# creating click task group
+@click.group(cls=AliasedGroup)
+def task():
+        pass
+
+
+# questionary function initialization
 def confirmation_question(question):
     answer = questionary.confirm(question, default=False).ask()
     return answer
@@ -43,7 +57,9 @@ def listing_categories(question, choices):
 
 
 def status_color(status):
+    # function to return emoji for given status
     color_status = ""
+
     # setting emoji for status
     if status == "TODO":
         color_status = "⏳"
@@ -55,6 +71,7 @@ def status_color(status):
 
 
 def get_status():
+    # list containing status with emoji
     status = [
                 'TODO ⏳',
                 'DONE ✅',
@@ -63,34 +80,12 @@ def get_status():
     return status
 
 
-@cli.command(name='new-category')
-def create_category():
-    """Create a new category """
-    # get value
-    category_name = input_question("Category name:")
-    category_emoji = input_question("Insert emoji:")
-
-    database = TinyDB(DB_FILE_CATEGORY)
-    database.insert({
-            'id': category_name,
-            'emoji': category_emoji
-        })
-    print(f"Created {category_name}")
-
-
-def get_all_category():
-    database = TinyDB(DB_FILE_CATEGORY)
-
-    categories = []
-    for category in database.all():
-        categories.append(f"{category['emoji']} {category['id']}")
-    return categories
-
-
 def get_category_emoji(category):
-    # just in case there is no category
+    # if no category given, return default emoji
     if category == None:
         return "📋"
+
+    # getting emoji for given category
     category = category[2:]
     database = TinyDB(DB_FILE_CATEGORY)
     Category = Query()
@@ -98,37 +93,35 @@ def get_category_emoji(category):
     return result[0]['emoji']
 
 
-@cli.command(name='list-categories')
-def list_category():
-    """List all categories"""
+def get_all_category():
+    # database initialization
     database = TinyDB(DB_FILE_CATEGORY)
-    categories = database.all()
 
-    console = Console()
-    table = Table(box=box.SIMPLE)
-    table.add_column('Category name')
-    table.add_column('Emoji')
-
-    for category in categories:
-        table.add_row(category['id'], category['emoji'])
-
-    console.print(table)
-
+    # creating a list and append all category to it
+    # then returning the said list
+    categories = []
+    for category in database.all():
+        categories.append(f"{category['emoji']} {category['id']}")
+    return categories
 
 
 def get_task_id(task):
+    # return task id
     return task.get('id')
 
 
 def generate_task_id():
+    # database initialization and fetching all tasks
     database = TinyDB(DB_FILE_TASKS)
     tasks_list = database.all()
 
+    # if no task is present, we assume it's first id
     if not tasks_list:
         return 'TASK-0001'
 
+    # sorting all task_id and returning new id based on 
+    # last highest one, plus one
     tasks_list.sort(key=get_task_id, reverse=True)
-    # getting returned value and iterating it
     highest_tasks = tasks_list[0]['id']
     number = int(highest_tasks[5:]) + 1
     missing_zeros = 4 - len(str(number))
@@ -136,7 +129,7 @@ def generate_task_id():
     return task_id
 
 
-@cli.command(name='new-task')
+@task.command(name='new')
 def create_task():
     """Create a new task"""
     status = "TODO"
@@ -161,7 +154,7 @@ def create_task():
                 not_empty = True
     else:
         description = None
-    
+
     # asking for a category
     need_category = confirmation_question("Is a category needed ?")
     if need_category:
@@ -169,7 +162,7 @@ def create_task():
     else:
         category = None
 
-    # asking for a due date 
+    # asking for a due date
     date_format = "%Y-%m-%d"
     need_duedate = confirmation_question("Is a due date required ?")
     if need_duedate:
@@ -177,7 +170,7 @@ def create_task():
         date_not_passed = False
         while date_format_valid is not True and date_not_passed is not True:
             try:
-                due_date = input_question("Enter due date (YYYY-MM-DD): ")
+                due_date = input_question("Enter due date (YYYY-MM-DD):")
                 datetime.strptime(due_date, date_format)
                 date_valid = True
             except ValueError:
@@ -197,7 +190,7 @@ def create_task():
     creation_date = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
     last_update = creation_date
 
-    # setting task_id 
+    # setting task_id
     task_id = generate_task_id()
 
     new_task = {
@@ -216,20 +209,24 @@ def create_task():
 
     # return the created task_id
     print("Created task: " + task_id)
+    return 0
 
 
-@cli.command()
+@task.command(name="get")
 @click.argument('task_id', required=True)
 def get_task(task_id):
     """Get task details"""
+
+    # database initialization
     database = TinyDB(DB_FILE_TASKS)
     Task = Query()
     result = database.search(Task.id == task_id)
 
-    # create table and display them
+    # rich table and console initialization
     console = Console()
     table = Table(box=box.SIMPLE)
 
+    # appending data to rich table
     table.add_column(f"{result[0]['id']}")
     table.add_row(f"Name : {result[0]['name']}")
     table.add_row(f"Status : {status_color(result[0]['status'])} ({result[0]['status']})")
@@ -238,27 +235,32 @@ def get_task(task_id):
         table.add_row(f"Due date : {result[0]['due_date']}")
     table.add_row(f"Creation date : {result[0]['creation_date']}")
     table.add_row(f"Last update : {result[0]['last_update']}")
+
+    # printing first table
     console.print(table)
 
+    # printing description table if description
     if result[0]['description'] != None:
         table_desc = Table(box=box.SIMPLE, show_header=True)
         table_desc.add_column('Description')
         table_desc.add_row(result[0]['description'])
         console.print(table_desc)
+    return 0
 
 
-@cli.command(name='list-tasks')
+@task.command(name='list')
 @click.option('--filter',
               type=click.Choice(['ALL', 'DONE', 'ABORTED', 'TODO']),
-              default='ALL')
+              default='TODO')
 def get_all_task(filter):
     """List all tasks"""
+
+    # database initialization
     database = TinyDB(DB_FILE_TASKS)
     result = database.all()
 
+    # rich console and table initialization
     console = Console()
-
-    # creating table and assigning value into it
     table = Table(show_header=True, box=box.SIMPLE)
     table.add_column('ID')
     table.add_column('Name')
@@ -268,6 +270,7 @@ def get_all_task(filter):
     table.add_column('Creation date')
     table.add_column('Last update')
 
+    # appending data to rich table
     for task in result:
         if task['status'] in filter or filter == 'ALL':
             # getting emoji for display purpose
@@ -280,7 +283,7 @@ def get_all_task(filter):
                 category = task['category']
             if task['due_date'] == None:
                 due_date = ""
-            else: 
+            else:
                 due_date = task['due_date']
             status = status_color(task['status']) + f" {task['status']}"
 
@@ -293,21 +296,42 @@ def get_all_task(filter):
                     task['creation_date'],
                     task['last_update'],
                 )
-    
+
+    # print obtain table and return 0
     console.print(table)
+    return 0
 
 
-@cli.command(name='delete-task')
+@task.command(name='delete')
 @click.argument('task_id', required=True)
-def delete_task(task_id):
+@click.option('-f', '--force', is_flag=True, default=False, help='Delete without asking for validation')
+def delete_task(task_id, force):
     """Delete a task"""
+
+    # database initialization
     database = TinyDB(DB_FILE_TASKS)
     Task = Query()
-    database.remove(Task.id == task_id)
-    print(f"Deleted {task_id}")
+
+    # deleting without asking if force flag
+    if force:
+        database.remove(Task.id == task_id)
+        print(f"Deleted {task_id}")
+        return 0
+
+    # asking for user validation
+    confirmation = confirmation_question(f"Do you really wan't to delete {task_id}")
+
+    # if confirmation then delete, otherwise no
+    if confirmation:
+        database.remove(Task.id == task_id)
+        print(f"Deleted {task_id}")
+        return 0
+    else:
+        print("Deletion cancelled")
+        return 0
 
 
-@cli.command(name='update-task')
+@task.command(name='update')
 @click.argument('task_id', required=True)
 def update_task(task_id):
     """Update a task"""
@@ -323,16 +347,7 @@ def update_task(task_id):
         if not description:
             description = None
         database.update({'description': description}, Task.id == task_id)
+    if need_update_status or need_update_description:
+        new_date_update = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+        database.update({'last_update': new_date_update}, Task.id == task_id)
 
-
-def first_initialization():
-    dir_path = os.environ['HOME'] + '/.pytodo'
-    if path.exists(dir_path):
-        return None
-    else:
-        os.mkdir(dir_path)
- 
-
-if __name__ == '__main__':
-    first_initialization()
-    cli()
